Move tooltip logic to separate file and export AA mapping modification
